+++ src/Objects/Events.php
@@ -7,13 +7,9 @@ use andrefelipe\Orchestrate\Query\TimeRangeBuilder;
 
 class Events extends AbstractList
 {
+    use EventClassTrait;
     use KeyTrait;
     use TypeTrait;
-
-    /**
-     * @var string
-     */
-    protected static $itemKind = 'event';
 
     /**
      * @param string $collection
@@ -27,6 +23,30 @@ class Events extends AbstractList
         $this->setType($type);
     }
 
+    /**
+     * @return int
+     */
+    public function getTotalCount()
+    {
+        if ($this->_totalCount === null) {
+
+            // makes a straight Search query for no results
+            $path = $this->getCollection(true);
+            $parameters = [
+                'query' => '@path.kind:event',
+                'limit' => 0,
+            ];
+            $response = $this->getHttpClient(true)->request('GET', $path, ['query' => $parameters]);
+
+            // set value if succesful
+            if ($response->getStatusCode() === 200) {
+                $body = $response->json();
+                $this->_totalCount = !empty($body['total_count']) ? (int) $body['total_count'] : 0;
+            }
+        }
+        return $this->_totalCount;
+    }
+
     public function reset()
     {
         parent::reset();
@@ -37,31 +57,36 @@ class Events extends AbstractList
     public function init(array $data)
     {
         if (!empty($data)) {
-            parent::init($data);
 
+            if (!empty($data['eventClass'])) {
+                $this->setEventClass($data['eventClass']);
+            }
             if (isset($data['key'])) {
                 $this->setKey($data['key']);
             }
             if (isset($data['type'])) {
                 $this->setType($data['type']);
             }
+
+            parent::init($data);
         }
         return $this;
     }
 
     public function toArray()
     {
-        $result = parent::toArray();
-        $result['kind'] = 'events';
+        $data = parent::toArray();
+        $data['kind'] = 'events';
+        $data['eventClass'] = $this->getEventClass()->name;
 
         if (!empty($this->_key)) {
-            $result['key'] = $this->_key;
+            $data['key'] = $this->_key;
         }
         if (!empty($this->_type)) {
-            $result['type'] = $this->_type;
+            $data['type'] = $this->_type;
         }
 
-        return $result;
+        return $data;
     }
 
     /**
@@ -90,5 +115,51 @@ class Events extends AbstractList
         $this->request('GET', $path, ['query' => $parameters]);
 
         return $this->isSuccess();
+    }
+
+    /**
+     * Constructs an event instance. An Event or a custom class you set with setEventClass().
+     *
+     * @param string $key
+     * @param string $type
+     * @param int $timestamp
+     * @param int $ordinal
+     *
+     * @return EventInterface
+     */
+    public function event($key = null, $type = null, $timestamp = null, $ordinal = null)
+    {
+        return $this->getEventClass()->newInstance()
+                    ->setCollection($this->getCollection(true))
+                    ->setKey($key)
+                    ->setType($type)
+                    ->setTimestamp($timestamp)
+                    ->setOrdinal($ordinal)
+                    ->setHttpClient($this->getHttpClient(true));
+    }
+
+    /**
+     * @param array $itemValues
+     */
+    protected function createInstance(array $itemValues)
+    {
+        if (!empty($itemValues['path']['kind'])) {
+            $kind = $itemValues['path']['kind'];
+
+            if ($kind === 'event') {
+                $class = $this->getEventClass();
+
+            } else {
+                return null;
+            }
+
+            $item = $class->newInstance()->init($itemValues);
+
+            if ($client = $this->getHttpClient()) {
+                $item->setHttpClient($client);
+            }
+            return $item;
+        }
+        return null;
     }
 }