+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## HEAD
 - *Work in progress to implement Events Search*
 - 'setChildClass' is split into 'setItemClass' and 'setEventClass' as now Collections can search both items and events.
-- The HTTP is split to its own class, simplifing much of our code base.
+- The HTTP client is split to its own class, simplifing much of our code base.
 - 'setClient' method is now 'setHttpClient' so it doesn't confuse with the Client class.
 - Improved totalCount method to actually make an API call if the total count is not know yet.
 - Several improvements throughout the library.